Revert combine process order
The compensation benchmark file has to be processed after skills as
otherwise the output gets corrupt. This fix doesn't deal with the root
cause of that issue (corruption shouldn't happen for any file order),
but we need to be able to generate correct documentation so the revert
is the quickest fix.